# docs/swagger.yaml
basePath: /
definitions:
  internal_handler.CreateMatchRequest:
    properties:
      away_team_id:
        type: string
      home_team_id:
        type: string
      match_date:
        type: string
      match_time:
        type: string
    required:
    - away_team_id
    - home_team_id
    - match_date
    - match_time
    type: object
...
  internal_handler.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorDetail'
      success:
        type: boolean
    type: object
...
  internal_handler.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
...
  internal_handler.RevertRequest:
    properties:
      target_version:
        type: integer
    required:
    - target_version
    type: object
  internal_handler.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  internal_handler.UpdateMatchStatusRequest:
    properties:
      status:
        type: string
      version:
        type: integer
    required:
    - status
    - version
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for managing amateur football teams, players, matches, and
    reports for Perusahaan XYZ.
  title: Kickbase Football Team Management API
  version: "1.0"
paths:
...
  /api/matches:
    get:
      description: Get a paginated list of matches with optional team, status, date
        filtering and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by team ID (home or away)
        in: query
        name: team_id
        type: string
      - description: Filter by status (scheduled, completed, cancelled, deferred)
        in: query
        name: status
        type: string
      - description: Filter matches starting from date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Filter matches up to date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - default: match_date
        description: Sort field (match_date, match_time, created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.PaginatedResponse'
      summary: List all matches
      tags:
      - Matches
    post:
      consumes:
      - application/json
      description: Create a match schedule between two teams
      parameters:
      - description: Match data
        in: body
        name: match
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateMatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Schedule a new match
      tags:
      - Matches
  /api/matches/{id}:
    get:
      description: Get a single match by its ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get a match by ID
      tags:
      - Matches
  /api/matches/{id}/history:
    get:
      description: Get the version history of a match
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
      summary: Get match history
      tags:
      - Matches
  /api/matches/{id}/revert:
    post:
      consumes:
      - application/json
      description: Revert a match to a specific previous version
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Target version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RevertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Revert match to previous version
      tags:
      - Matches
  /api/matches/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a match (scheduled, completed, cancelled,
        deferred)
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateMatchStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update match status
      tags:
      - Matches
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"